Don't build any btest artifacts if --disable-btest is given
Some Btest tooling under btest/data and btest/scripts is still built if --disable-btest and --disable-btest-pcaps is passed to configure.

This commit adds a check in CMakeList file to ensure that does not happen. It also modifies the configure help to state that --disable-btest-pcaps is redundant if --disable-btest is passed