Clean up fully after successful Docker btests
If we leave files sitting around, we trigger a Docker image double-build in CI,
because the build runs once, gets tested, and then gets run again when we push
the Docker image: the additional btest files cause Docker to detect a different
source tree, causing an image layer violation.

Also rename "cleanup" target to "clean", to align with rest of our tree.