Fix document generation
We broke multi-repo git interaction when we introduced use of "actions/checkout",
because it has subtle implications on auth mechanisms (for details, check action
logs when it conducts a recursive checkout). Thankfully it also helps overcome
these: passing a user token to actions/checkout simplifies multi-repo git interaction,
making it work as it should: you no longer need to think about it. We now use this
approach.

Also tweak the git commit logic: we no longer mask all errors. Seems we mainly
did that to mask "git commit" on no changes erroring out. Instead we now check
whether there's anything to commit, and only do so then. We do this both in the
zeek-docs repo and for the docs submodule bump.