Initial implementation of Count-Min Sketch
Count-Min sketch is a probabilistic data structure somewhat similar to
counting bloom filters, with a bit different characteristics. One of the
main characteristics is that it only overestimates.

This is a very rough implementation that just follows the paper. It is
barely functional and missing many niceties - like support for cluster,
or stable tests.