Fix SSL tests on Centos Stream 9
SHA1 signatures are only allowed in the legacy policy. I have no idea
where this is documented - I found it in a patch for the OpenSSL source
package.

Also adding tar as a specific dependency, since we have a test that
requires it, and at least for me it was not present in a minimal
installation.

Fixes #2115